apps/worker/requirements.txt
chess==1.11.2
stockfish==3.28.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
